src/main/java/com/example/journalservice/controller/UserEventController.java
package com.example.journalservice.controller;

import com.example.journalservice.dto.UserEventResponseDto;
import com.example.journalservice.service.UserEventService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequiredArgsConstructor
@RequestMapping("/api/v1")
public class UserEventController {

    private final UserEventService userEventService;

    @PreAuthorize("hasAuthority('ADMIN')")
    @GetMapping("/all")
    public ResponseEntity<List<UserEventResponseDto>> getAllUserEvents() {
        return ResponseEntity.ok(userEventService.getAllUserEvents());
    }

}
